refactor(pdf-generators): reorganize types and interfaces for document services
