+++ packages/pdf-generators/src/modules/pdf_generators/services/generation-history-service/generation-history-service.ts
@@ -1,46 +1,10 @@
 import type { EntityManager, FilterQuery } from '@mikro-orm/postgresql'
 import { GeneratedDocument } from '../../data/entities'
-
-/** Organization + tenant pair every history operation is scoped to. */
-export interface HistoryScope {
-  organizationId: string
-  tenantId: string
-}
-
-/** Input for persisting a single generation-history row. */
-export interface CreateGeneratedDocumentInput {
-  scope: HistoryScope
-  templateId: string
-  templateLabel: string
-  resourceKind: string
-  resourceId: string
-  resourceLabel: string
-  generatedBy: string
-  format?: string
-  mimeType?: string
-}
-
-/** Input for a paginated history read. */
-export interface ListGeneratedDocumentsQuery {
-  scope: HistoryScope
-  page: number
-  pageSize: number
-  resourceKind?: string
-  resourceId?: string
-}
-
-/** Serialized history row returned to API callers (dates as ISO strings). */
-export interface GeneratedDocumentDto {
-  id: string
-  resourceKind: string
-  resourceId: string
-  resourceLabel: string
-  templateId: string
-  templateLabel: string
-  format: string
-  generatedBy: string
-  generatedAt: string
-}
+import type {
+  CreateGeneratedDocumentInput,
+  GeneratedDocumentDto,
+  ListGeneratedDocumentsQuery,
+} from './types'
 
 /**
  * Owns all persistence and read logic for {@link GeneratedDocument} history rows.
