+++ packages/pdf-generators/src/modules/pdf_generators/services/generation-history-service/index.ts
@@ -4,4 +4,4 @@ export type {
   CreateGeneratedDocumentInput,
   ListGeneratedDocumentsQuery,
   GeneratedDocumentDto,
-} from './generation-history-service'
+} from './types'
