+++ packages/pdf-generators/src/modules/pdf_generators/services/orders-document-service/index.ts
@@ -1,4 +1,4 @@
 export { OrdersDocumentService, ORDERS_TEMPLATE_IDS } from './orders-document-service'
-export type { OrderRecord, OrderLineItem } from './orders-document-service'
+export type { OrderRecord, OrderLineItem } from './types'
 export { orderDocumentInputSchema } from './validators'
 export type { OrderDocumentInput } from './validators'
