+++ packages/pdf-generators/src/modules/pdf_generators/services/quotes-document-service/index.ts
@@ -1,4 +1,4 @@
 export { QuotesDocumentService, QUOTES_TEMPLATE_IDS } from './quotes-document-service'
-export type { QuoteRecord, QuoteLineItem } from './quotes-document-service'
+export type { QuoteRecord, QuoteLineItem } from './types'
 export { quoteDocumentInputSchema } from './validators'
 export type { QuoteDocumentInput } from './validators'
